Extract Jp2Creator class
