+++ .rubocop.yml
@@ -4,6 +4,7 @@ require: rubocop-rspec
 
 AllCops:
   TargetRubyVersion: 3.0
+  SuggestExtensions: false
 
 # rspec expect{...} is conventional
 Layout/SpaceBeforeBlockBraces:
